python/requirements.txt
google-genai
pyaudio
opencv-python
python-dotenv
websockets
